Add check for 'print-pdf' query string
If we find 'print-pdf' then we'll have to remove it from the request
url. I've tried a few ways to go about this but in the end this was
the simplest change, maybe not the most elegant. I think this way is
the closest to what reveal.js already does.

I tried:
* routing - adding a route that supports the keyword print
  (eg /print/MARKDOWN_REGEX)
* query - directly checking for query string in the server request
  (eg req.query, but this was always stripped out by the route)