Add support for print pdf from the command line
Yet another wrapper feature from reveal.js.
This leverages the reveal.js print-pdf plugin by passing the server
the `--print` (or `-r`) parameter.

It will try to build the pdf based on the initial file that the server
loads. So if no file is passed to the server the print-pdf will build
a pdf for the server listing, ie `reveal-md --print blah.pdf` will
produce a pdf of the whole server listing of markdown files.