Make ?print-pdf actually work.
Current code in options.js doesn't work because the 'query.print-pdf' property
will have an empty string as its value when the URL is like '?print-pdf',
evaluating to false.

Even when it evaluates to true, it'll be a problem because then reveal-md
handles it like the --print command-line flag has been passed and tries to
invoke phantomjs.

Instead, sets 'view.print' right before passing it to mustache.