Insert OpenGraph metadata to html template
OpenGraph metadata allows nice rendering output of links to static
builds when they are shared in social media sites.

Metada is generated only when ``--absolute-url`` command line option
is used to define eventual hosted url of the slide deck. This is used
to create canonical url and ``og:image`` url.