install dependencies before the app
it is a best practice to install dependencies first because the prior
layers are not rebuilt (cache not invalidated) if the depdencies do not
change.

since application code tends to change more often, it should be in the
higher layers (ordered last in the Dockerfile).

add package-lock.json to source control, as advised by npm.