Fix copying every images that looks like <img> tag
Instead of doing regex search on markdown source,
(both markdown `![alt](/img/src/path "title")]` and html `<img src="/path" />`)
convert it to html first and then search only for html.

This fixes the problem of falsely trying to copy img tags that are not
actual images, but codes:

```html
<img src="please/dont/bother/copying" />
```

which is HTML-escaped and is safe from being caught by `htmlImageRe`.